+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - introducing `withLock` wrapper
 - improved error messages by using command name prefix (fix #88)
+- multiline pipelines (fix #83)
 
 ### Fixed
 - allowing sequence commands everywhere (fix #130)